+++ src/lib/lua_mympd_state.h
@@ -27,13 +27,13 @@ enum lua_mympd_state_type {
  */
 struct t_lua_mympd_state_value {
     sds p;        //!< lua string
-    long long i;  //!< lua number
+    int64_t i;  //!< lua number
     double f;     //!< lua float
     bool b;       //!< lua bool
 };
 
 void lua_mympd_state_set_p(struct t_list *lua_mympd_state, const char *k, const char *v);
-void lua_mympd_state_set_i(struct t_list *lua_mympd_state, const char *k, long long v);
+void lua_mympd_state_set_i(struct t_list *lua_mympd_state, const char *k, int64_t v);
 void lua_mympd_state_set_u(struct t_list *lua_mympd_state, const char *k, unsigned v);
 void lua_mympd_state_set_b(struct t_list *lua_mympd_state, const char *k, bool v);
 void lua_mympd_state_set_f(struct t_list *lua_mympd_state, const char *k, double v);
