+++ src/mpd_client/random_select.h
@@ -19,10 +19,10 @@ struct t_random_add_constraints {
     unsigned max_song_duration;
 };
 
-long random_select_albums(struct t_partition_state *partition_state, struct t_stickerdb_state *stickerdb,
-        struct t_cache *album_cache, long add_albums, struct t_list *queue_list, struct t_list *add_list,
+unsigned random_select_albums(struct t_partition_state *partition_state, struct t_stickerdb_state *stickerdb,
+        struct t_cache *album_cache, unsigned add_albums, struct t_list *queue_list, struct t_list *add_list,
         struct t_random_add_constraints *constraints);
-long random_select_songs(struct t_partition_state *partition_state, struct t_stickerdb_state *stickerdb,
-        long add_songs, const char *playlist, struct t_list *queue_list, struct t_list *add_list,
+unsigned random_select_songs(struct t_partition_state *partition_state, struct t_stickerdb_state *stickerdb,
+        unsigned add_songs, const char *playlist, struct t_list *queue_list, struct t_list *add_list,
         struct t_random_add_constraints *constraints);
 #endif
