+++ src/mympd_api/browse.h
@@ -10,10 +10,10 @@
 #include "src/lib/mympd_state.h"
 
 sds mympd_api_browse_album_detail(struct t_mympd_state *mympd_state, struct t_partition_state *partition_state,
-        sds buffer, long request_id, sds albumid, const struct t_tags *tagcols);
+        sds buffer, unsigned request_id, sds albumid, const struct t_tags *tagcols);
 sds mympd_api_browse_album_list(struct t_partition_state *partition_state, struct t_cache *album_cache,
-        sds buffer, long request_id, sds expression, sds sort, bool sortdesc, long offset, long limit,
+        sds buffer, unsigned request_id, sds expression, sds sort, bool sortdesc, unsigned offset, unsigned limit,
         const struct t_tags *tagcols);
 sds mympd_api_browse_tag_list(struct t_partition_state *partition_state, sds buffer,
-        long request_id, sds searchstr, sds tag, long offset, long limit, bool sortdesc);
+        unsigned request_id, sds searchstr, sds tag, unsigned offset, unsigned limit, bool sortdesc);
 #endif
