+++ src/mympd_api/scripts.h
@@ -16,8 +16,8 @@
     bool mympd_api_script_save(sds workdir, sds script, sds oldscript, int order, sds content, struct t_list *arguments, sds *error);
     bool mympd_api_script_validate(sds name, sds content, sds lualibs, sds *error);
     bool mympd_api_script_delete(sds workdir, sds script);
-    sds mympd_api_script_get(sds workdir, sds buffer, long request_id, sds script);
-    sds mympd_api_script_list(sds workdir, sds buffer, long request_id, bool all);
+    sds mympd_api_script_get(sds workdir, sds buffer, unsigned request_id, sds script);
+    sds mympd_api_script_list(sds workdir, sds buffer, unsigned request_id, bool all);
     bool mympd_api_script_start(sds workdir, sds script, sds lualibs, struct t_list *arguments,
             const char *partition, bool localscript);
 #endif
