+++ src/mympd_api/volume.c
@@ -26,7 +26,7 @@
  * @return pointer to buffer
  */
 sds mympd_api_volume_set(struct t_partition_state *partition_state, unsigned volume_min, unsigned volume_max,
-        sds buffer, enum mympd_cmd_ids cmd_id, long request_id, unsigned volume)
+        sds buffer, enum mympd_cmd_ids cmd_id, unsigned request_id, unsigned volume)
 {
     if (volume > volume_max ||
         volume < volume_min)
@@ -56,7 +56,7 @@ sds mympd_api_volume_set(struct t_partition_state *partition_state, unsigned vol
  * @return pointer to buffer
  */
 sds mympd_api_volume_change(struct t_partition_state *partition_state, unsigned volume_min, unsigned volume_max,
-        sds buffer, long request_id, int relative_volume)
+        sds buffer, unsigned request_id, int relative_volume)
 {
     //calculate absolute volume
     int curVolume = mpd_client_get_volume(partition_state);
