+++ src/mympd_api/volume.h
@@ -12,8 +12,8 @@
 #include "src/lib/mympd_state.h"
 
 sds mympd_api_volume_set(struct t_partition_state *partition_state, unsigned volume_min, unsigned volume_max,
-        sds buffer, enum mympd_cmd_ids cmd_id, long request_id, unsigned volume);
+        sds buffer, enum mympd_cmd_ids cmd_id, unsigned request_id, unsigned volume);
 sds mympd_api_volume_change(struct t_partition_state *partition_state, unsigned volume_min, unsigned volume_max,
-        sds buffer, long request_id, int relative_volume);
+        sds buffer, unsigned request_id, int relative_volume);
 
 #endif
